refactor

+++ frame/system/module.go
@@ -68,7 +68,7 @@ func (m Module[N]) Initialize(blockNumber N, parentHash primitives.Blake2bHash, 
 	m.Storage.BlockNumber.Put(blockNumber)
 	m.Storage.Digest.Put(digest)
 	m.Storage.ParentHash.Put(parentHash)
-	m.Storage.BlockHash.Put(blockNumber.Sub(sc.NewNumeric[N](uint8(1))).(N), parentHash)
+	m.Storage.BlockHash.Put(blockNumber.Sub(sc.NewNumeric[N](1)).(N), parentHash)
 	m.Storage.BlockWeight.Clear()
 }
 
@@ -157,14 +157,14 @@ func (m Module[N]) Finalize() primitives.Header[N] {
 	buf.Reset()
 
 	// saturating_sub
-	v := sc.NewNumeric[N](uint32(m.Constants.BlockHashCount)).Sub(sc.NewNumeric[N](uint8(1)))
+	v := sc.NewNumeric[N](m.Constants.BlockHashCount).Sub(sc.NewNumeric[N](1))
 	toRemove := blockNumber.Sub(v)
 
 	if toRemove.Gt(blockNumber) {
-		toRemove = sc.NewNumeric[N](uint8(0))
+		toRemove = sc.NewNumeric[N](0)
 	}
 
-	if toRemove.Ne(sc.NewNumeric[N](uint8(0))) {
+	if toRemove.Ne(sc.NewNumeric[N](0)) {
 		m.Storage.BlockHash.Remove(toRemove.(N))
 	}
 
@@ -341,7 +341,7 @@ func (m Module[N]) decProviders(who primitives.Address32) (primitives.DecRefStat
 // NOTE: Events not registered at the genesis block and quietly omitted.
 func (m Module[N]) depositEventIndexed(topics []primitives.H256, event primitives.Event) {
 	blockNumber := m.Storage.BlockNumber.Get()
-	if blockNumber.Eq(sc.NewNumeric[N](uint8(0))) {
+	if blockNumber.Eq(sc.NewNumeric[N](0)) {
 		return
 	}
